+++ src/channels/category.rs
@@ -1,11 +1,14 @@
-// Copyright (c) 2016 Chris Palmer <pennstate5013@gmail.com>
+// Copyright (c) 2015-2016 Chris Palmer <pennstate5013@gmail.com>
 // Use of this source code is governed by the LGPLv3 license that can be
 // found in the LICENSE file.
+
 
 //! The fields under category can be retrieved by using the methods under
 //! `Category`.
 
-use rss::Category;
+
+use channels::Category;
+
 
 impl Category {
     /// Get the category that exists under `Category`.
@@ -13,7 +16,7 @@ impl Category {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::CategoryBuilder;
+    /// use feed::channels::CategoryBuilder;
     ///
     /// let category = "podcast";
     /// let category_obj = CategoryBuilder::new()
@@ -31,7 +34,7 @@ impl Category {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::CategoryBuilder;
+    /// use feed::channels::CategoryBuilder;
     ///
     /// let domain_string = "http://jupiterbroadcasting.com".to_owned();
     /// let category = CategoryBuilder::new()
@@ -44,7 +47,7 @@ impl Category {
     /// ```
     ///
     /// ```
-    /// use feed::rss::CategoryBuilder;
+    /// use feed::channels::CategoryBuilder;
     ///
     /// let category = CategoryBuilder::new()
     ///     .domain(None)