+++ src/channels/guid.rs
@@ -1,10 +1,13 @@
-// Copyright (c) 2016 Chris Palmer <pennstate5013@gmail.com>
+// Copyright (c) 2015-2016 Chris Palmer <pennstate5013@gmail.com>
 // Use of this source code is governed by the LGPLv3 license that can be
 // found in the LICENSE file.
 
+
 //! The fields under guid can be retrieved by using the methods under `Guid`.
 
-use rss::Guid;
+
+use channels::Guid;
+
 
 impl Guid {
     /// Get the permalink that exists under `Guid`.
@@ -12,7 +15,7 @@ impl Guid {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::GuidBuilder;
+    /// use feed::channels::GuidBuilder;
     ///
     /// let guid = GuidBuilder::new()
     ///     .permalink(None)
@@ -21,7 +24,7 @@ impl Guid {
     /// ```
     ///
     /// ```
-    /// use feed::rss::GuidBuilder;
+    /// use feed::channels::GuidBuilder;
     ///
     /// let permalink = true;
     /// let guid = GuidBuilder::new()
@@ -31,7 +34,7 @@ impl Guid {
     /// ```
     ///
     /// ```
-    /// use feed::rss::GuidBuilder;
+    /// use feed::channels::GuidBuilder;
     ///
     /// let permalink = false;
     /// let guid = GuidBuilder::new()
@@ -49,7 +52,7 @@ impl Guid {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::GuidBuilder;
+    /// use feed::channels::GuidBuilder;
     ///
     /// let guid = "9DE46946-2F90-4D5D-9047-7E9165C16E7C";
     /// let guid_obj = GuidBuilder::new()