+++ src/channels/source.rs
@@ -1,10 +1,13 @@
-// Copyright (c) 2016 Chris Palmer <pennstate5013@gmail.com>
+// Copyright (c) 2015-2016 Chris Palmer <pennstate5013@gmail.com>
 // Use of this source code is governed by the LGPLv3 license that can be
 // found in the LICENSE file.
 
+
 //! The fields under source can be retrieved by using the methods under `Source`.
 
-use rss::Source;
+
+use channels::Source;
+
 
 impl Source {
     /// Get the url that exists under `Source`.
@@ -12,7 +15,7 @@ impl Source {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::SourceBuilder;
+    /// use feed::channels::SourceBuilder;
     ///
     /// let url = "http://www.tomalak.org/links2.xml";
     /// let source = SourceBuilder::new()
@@ -30,7 +33,7 @@ impl Source {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::SourceBuilder;
+    /// use feed::channels::SourceBuilder;
     ///
     /// let source = "Tomalak's Realm";
     /// let source_obj = SourceBuilder::new()