+++ src/channels/source_builder.rs
@@ -1,10 +1,13 @@
-// Copyright (c) 2016 Chris Palmer <pennstate5013@gmail.com>
+// Copyright (c) 2015-2016 Chris Palmer <pennstate5013@gmail.com>
 // Use of this source code is governed by the LGPLv3 license that can be
 // found in the LICENSE file.
 
+
 //! The fields can be set for source by using the methods under `SourceBuilder`.
 
-use rss::{Source, SourceBuilder};
+
+use channels::{Source, SourceBuilder};
+
 
 impl SourceBuilder {
     /// Construct a new `SourceBuilder` and return default values.
@@ -12,7 +15,7 @@ impl SourceBuilder {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::SourceBuilder;
+    /// use feed::channels::SourceBuilder;
     ///
     /// let source_builder = SourceBuilder::new();
     /// ```
@@ -26,7 +29,7 @@ impl SourceBuilder {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::SourceBuilder;
+    /// use feed::channels::SourceBuilder;
     ///
     /// let mut source_builder = SourceBuilder::new();
     /// source_builder.url("http://www.example.com/source");
@@ -42,7 +45,7 @@ impl SourceBuilder {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::SourceBuilder;
+    /// use feed::channels::SourceBuilder;
     ///
     /// let mut source_builder = SourceBuilder::new();
     /// source_builder.source("Test");
@@ -58,7 +61,7 @@ impl SourceBuilder {
     /// # Examples
     ///
     /// ```
-    /// use feed::rss::SourceBuilder;
+    /// use feed::channels::SourceBuilder;
     ///
     /// let source = SourceBuilder::new()
     ///         .url("http://www.example.com/source")